Extract some funcs to `utils.py`
