+++ src/san_strings/__init__.py
@@ -1,63 +1,11 @@
-import math
-from functools import cache
 from typing import Literal
 
 import chess
 
+from san_strings.utils import extend_ray
+
 _b = chess.Board.empty()
 """ A blank `chess.Board` to use for generating moves. """
-
-
-@cache
-def _sliding_delta(s1: chess.Square, s2: chess.Square) -> int:
-    """
-    Get the delta of the index in `chess.SQUARES` required to move
-    one step from `s1` toward `s2`. Raises `AssertionError` if `s1`
-    and `s2` are not on the same file, rank, or diagonal.
-
-    >>> _sliding_delta(chess.C3, chess.F6)
-    9
-    >>> chess.C3 + 9 == chess.D4
-    True
-    """
-    assert s1 != s2, 's1 and s2 must be different squares'
-
-    x_delta = chess.square_file(s2) - chess.square_file(s1)
-    y_delta = chess.square_rank(s2) - chess.square_rank(s1)
-
-    assert 0 in (x_delta, y_delta) or abs(x_delta) == abs(y_delta), (
-        's1 and s2 must be on the same file, rank, or diagonal; got '
-        f'{chess.square_name(s1)} and {chess.square_name(s2)}'
-    )
-
-    x_delta = int(math.copysign(1, x_delta))
-    y_delta = int(math.copysign(1, y_delta))
-
-    return y_delta * 8 + x_delta
-
-
-@cache
-def _extend_ray(
-    from_sq: chess.Square,
-    towards_sq: chess.Square,
-) -> chess.Bitboard:
-    """
-    Get a `chess.Bitboard` of all the squares from (and including) `from_sq`
-    toward `towards_sq` and continuing on to an edge of the board.
-
-    >>> print(chess.SquareSet(_extend_ray(chess.C3, chess.F6)))
-    . . . . . . . 1
-    . . . . . . 1 .
-    . . . . . 1 . .
-    . . . . 1 . . .
-    . . . 1 . . . .
-    . . 1 . . . . .
-    . . . . . . . .
-    . . . . . . . .
-    """
-    d = _sliding_delta(from_sq, towards_sq)
-    # noinspection PyProtectedMember
-    return chess._sliding_attacks(from_sq, 0, [d]) | chess.BB_SQUARES[from_sq]
 
 
 def get_pawn_sans(only_for_color: chess.Color | None = None) -> set[str]:
@@ -216,7 +164,7 @@ def get_piece_sans(symbol: Literal['N', 'B', 'R', 'Q']) -> set[str]:
             from_square_rank = chess.square_rank(from_square)
 
             if is_sliding_piece:
-                bb_ray = _extend_ray(to_square, from_square)
+                bb_ray = extend_ray(to_square, from_square)
 
             bb_from_square_file = chess.BB_FILES[from_square_file]
             bb_from_square_rank = chess.BB_RANKS[from_square_rank]
